Use semantic versioning
Fixes #10

When trying to import this via NPM (slowly becoming the standard way to do so with Hubot), NPM complains about the package not using [semantic versioning](http://semver.org/) and halts installation. Ths PR fixes that issue, as well as renames the NPM name so that it lines up with the package name itself.