🔧 "redmine search" but not "search redmine"
If we respond to "search redmine", we'll also be occupying the same
space as other commands like [hubot-search](https://www.npmjs.com/package/hubot-search), which means
that hubot will respond to *both* commands, and the "search" command
will be "redmine searchstring".

So to avoid conflict / unexpected behaviour, I'll just stick with
"redmine search" for now.