Refactor to use node-redmine, enhance test suite
Fixes #33

This started out as an attempt to add tests to cover existing
functionality, but I ran into a fair amount of trouble with the wrapper
that was added to talk to the redmine install (kept duplicating calls
when under test). I finally opted to dig around for a way to abstract
that requirement into another module (`node-redmine`). After that was
complete, I got some minimally viable test coverage on each of the
commands.

Some notes:

- Removed some instance-specific code around issue statuses
- Removed where a particular user ID would get a sad trombone played for them when an issue was assigned to them
- General cleanup