Namespace commands, default issue listener to off
Prior to this PR, users may run into command collisions with other Hubot
packages. Additionally, the default listener for replying with issue
often matched too many messages in a typical chat room. This PR
namespaces the commands to use either `redmine` or the shorter `rm` and
requires the user to set a custom listener `HUBOT_REDMINE_MENTION_REGEX`
to enable that listener to be triggered.

Cleans up documentation in README.

Fixes #21